refactor(emit): split es6 class emitter phases
