+++ crates/tsz-emitter/src/emitter/declarations/class/emit_es6.rs
@@ -1,109 +1,31 @@
-use super::super::super::core::PropertyNameEmit;
 use super::super::super::{Printer, ScriptTarget};
+use super::AutoAccessorInfo;
 use super::duplicate_private_names::{
     PrivateDuplicateConflictPlan, collect_private_duplicate_conflicts,
 };
-use super::private_comma_items::PrivateCommaItems;
-use super::replace_identifier;
-use super::static_field_erasure::static_no_init_field_is_erased;
-use super::{AutoAccessorEmitOptions, AutoAccessorInfo, StaticFieldInit};
+use super::emit_es6_after_body::ClassEs6AfterBody;
+use super::emit_es6_field_inits::ClassFieldInitCollection;
+use super::emit_es6_members::ClassEs6MemberEmit;
+use super::emit_es6_private_accessors::{
+    PrivateAutoAccessorInfo, collect_private_auto_accessors_with_reserved,
+};
 use crate::emitter::core::{
-    PrivateAccessorDef, PrivateFieldStorageKind, PrivateMemberInfo, PrivateMethodDef,
-    StaticPrivateInit,
+    PrivateFieldStorageKind, PrivateMemberInfo, PrivateMethodDef, StaticPrivateInit,
 };
 use crate::transforms::private_fields_es5::{
     PrivateAccessorInfo, PrivateFieldInfo, PrivateMethodInfo,
     collect_enclosing_source_binding_names, collect_private_members_with_reserved,
     get_private_field_name, is_private_identifier, make_unique_private_name, private_helper_base,
 };
-use rustc_hash::{FxHashMap, FxHashSet};
+use rustc_hash::FxHashMap;
 use std::sync::Arc;
 use tsz_parser::parser::NodeIndex;
-use tsz_parser::parser::node::{ClassData, Node, NodeAccess};
+use tsz_parser::parser::node::{Node, NodeAccess};
 use tsz_parser::parser::syntax_kind_ext;
 use tsz_parser::syntax::transform_utils::{
     contains_async_arrow_function, contains_super_reference, contains_this_reference,
 };
 use tsz_scanner::SyntaxKind;
-
-#[derive(Debug, Clone)]
-pub(super) struct PrivateAutoAccessorInfo {
-    pub(super) member_idx: NodeIndex,
-    pub(super) name: String,
-    pub(super) get_var_name: String,
-    pub(super) set_var_name: String,
-    pub(super) storage_name: String,
-    pub(super) initializer: Option<NodeIndex>,
-    pub(super) is_static: bool,
-}
-
-fn collect_private_auto_accessors_with_reserved(
-    printer: &Printer<'_>,
-    class: &ClassData,
-    class_name: &str,
-    used_names: &mut rustc_hash::FxHashSet<String>,
-) -> Vec<PrivateAutoAccessorInfo> {
-    if class_name.is_empty() {
-        return Vec::new();
-    }
-
-    let mut accessors = Vec::new();
-    for &member_idx in &class.members.nodes {
-        let Some(member_node) = printer.arena.get(member_idx) else {
-            continue;
-        };
-        let Some(prop) = printer.arena.get_property_decl(member_node) else {
-            continue;
-        };
-        if !printer
-            .arena
-            .has_modifier(&prop.modifiers, SyntaxKind::AccessorKeyword)
-        {
-            continue;
-        }
-        if printer
-            .arena
-            .has_modifier(&prop.modifiers, SyntaxKind::AbstractKeyword)
-            || printer
-                .arena
-                .has_modifier(&prop.modifiers, SyntaxKind::DeclareKeyword)
-        {
-            continue;
-        }
-        if !is_private_identifier(printer.arena, prop.name) {
-            continue;
-        }
-
-        let Some(field_name) = get_private_field_name(printer.arena, prop.name) else {
-            continue;
-        };
-        let clean_name = field_name.strip_prefix('#').unwrap_or(&field_name);
-        let base = format!("_{class_name}_{clean_name}");
-        let get_var_name = make_unique_private_name(&format!("{base}_get"), used_names);
-        let set_var_name = make_unique_private_name(&format!("{base}_set"), used_names);
-        let storage_name = if printer.ctx.options.legacy_decorators {
-            used_names.insert(base.clone());
-            let storage_stem = make_unique_private_name(&base, used_names);
-            make_unique_private_name(&format!("{storage_stem}_accessor_storage"), used_names)
-        } else {
-            make_unique_private_name(&format!("{base}_accessor_storage"), used_names)
-        };
-        accessors.push(PrivateAutoAccessorInfo {
-            member_idx,
-            name: clean_name.to_string(),
-            get_var_name,
-            set_var_name,
-            storage_name,
-            initializer: if prop.initializer.is_none() {
-                None
-            } else {
-                Some(prop.initializer)
-            },
-            is_static: printer.has_effective_static_modifier_js(&prop.modifiers),
-        });
-    }
-    accessors
-}
 
 impl<'a> Printer<'a> {
     pub(in crate::emitter) fn class_expression_is_in_loop_body(&self, idx: NodeIndex) -> bool {
@@ -147,84 +69,6 @@ impl<'a> Printer<'a> {
 
     fn is_reserved_private_constructor_name(name: &str) -> bool {
         name == "constructor"
-    }
-
-    pub(super) fn emit_static_private_init(
-        &mut self,
-        init: &StaticPrivateInit,
-        class_name: &str,
-        with_semicolon: bool,
-    ) {
-        self.write(&init.storage_name);
-        match init.storage_kind {
-            PrivateFieldStorageKind::WeakMap => {
-                self.write(".set(");
-                self.write(class_name);
-                self.write(", ");
-            }
-            PrivateFieldStorageKind::Value => {
-                self.write(" = { value: ");
-            }
-        }
-        if init.initializer.is_some() {
-            self.emit_expression(init.initializer);
-        } else {
-            self.write("void 0");
-        }
-        match init.storage_kind {
-            PrivateFieldStorageKind::WeakMap => self.write(")"),
-            PrivateFieldStorageKind::Value => self.write(" }"),
-        }
-        if with_semicolon {
-            self.write(";");
-        }
-    }
-
-    pub(in crate::emitter) fn emit_private_auto_accessor_function_def(
-        &mut self,
-        var_name: &str,
-        storage_name: &str,
-        is_static: bool,
-        is_get: bool,
-        class_alias: Option<&str>,
-    ) {
-        self.write(var_name);
-        self.write(" = function ");
-        self.write(var_name);
-        self.write("(");
-        if !is_get {
-            self.write("value");
-        }
-        self.write(") { ");
-        if is_get {
-            self.write("return ");
-            self.write_helper("__classPrivateFieldGet");
-        } else {
-            self.write_helper("__classPrivateFieldSet");
-        }
-        self.write("(");
-        if is_static {
-            let alias = class_alias.unwrap_or("this");
-            self.write(alias);
-            self.write(", ");
-            self.write(alias);
-            if is_get {
-                self.write(", \"f\", ");
-                self.write(storage_name);
-            } else {
-                self.write(", value, \"f\", ");
-                self.write(storage_name);
-            }
-        } else {
-            self.write("this, ");
-            self.write(storage_name);
-            if is_get {
-                self.write(", \"f\"");
-            } else {
-                self.write(", value, \"f\"");
-            }
-        }
-        self.write("); }");
     }
 
     /// Emit a class using ES6 native class syntax (no transforms).
@@ -1722,131 +1566,13 @@ impl<'a> Printer<'a> {
                 static_super_base_alias.as_deref()
             };
 
-        if self.should_preserve_native_decorator_comments(&class.modifiers)
-            && let Some(name_node) = self.arena.get(class.name)
-        {
-            self.emit_comments_before_pos(name_node.pos);
-        }
-        self.write("class");
-
-        // Determine the class expression name.
-        // When assignment_prefix is provided (e.g., `let C = class C {}`), a named class
-        // keeps its name on the expression, but an anonymous class stays anonymous
-        // (`let default_1 = class {}`), even if anonymous_default_export_name is set.
-        if class.name.is_some() {
-            self.write_space();
-            self.emit_decl_name(class.name);
-        } else if assignment_prefix.is_none() {
-            // No assignment prefix — use anonymous_default_export_name if available
-            // (e.g., `export default class {}` → `class default_1 {}`)
-            let override_name = self.anonymous_default_export_name.clone();
-            if let Some(name) = override_name
-                && !name.is_empty()
-            {
-                self.write_space();
-                self.write(&name);
-            }
-        }
-
-        if let Some(ref heritage_clauses) = class.heritage_clauses {
-            for &clause_idx in &heritage_clauses.nodes {
-                let Some(clause_node) = self.arena.get(clause_idx) else {
-                    continue;
-                };
-                let Some(heritage) = self.arena.get_heritage(clause_node) else {
-                    continue;
-                };
-                if heritage.token != SyntaxKind::ExtendsKeyword as u16 {
-                    continue;
-                }
-
-                if !heritage.types.nodes.is_empty() {
-                    self.write(" extends ");
-                    for (i, &extends_type) in heritage.types.nodes.iter().enumerate() {
-                        if i > 0 {
-                            self.write(", ");
-                        }
-                        if let Some(base_alias) = static_super_base_alias.as_ref() {
-                            self.write("(");
-                            self.write(base_alias);
-                            self.write(" = ");
-                            self.emit_heritage_expression(extends_type);
-                            self.write(")");
-                        } else {
-                            self.emit_heritage_expression(extends_type);
-                        }
-                    }
-                } else {
-                    // Error recovery: source has `extends` with no base type.
-                    // Preserve the keyword to match tsc output.
-                    self.write(" extends ");
-                }
-            }
-        }
-
-        self.write(" {");
-        // Suppress trailing comments on class body opening brace.
-        // tsc drops same-line comments on `{` for class bodies, just like function
-        // bodies (e.g. `class C { // error` → `class C {`).
-        if !self.ctx.options.remove_comments
-            && let Some(text) = self.source_text
-        {
-            let bytes = text.as_bytes();
-            let start = node.pos as usize;
-            let end = (node.end as usize).min(bytes.len());
-            if let Some(offset) = bytes[start..end].iter().position(|&b| b == b'{') {
-                let brace_end = (start + offset + 1) as u32;
-                // Only suppress if there's a newline between `{` and the first
-                // member (or the closing `}` if empty).  Single-line class bodies
-                // like `class C { x: T; } // error` have the comment after `}`,
-                // so we must NOT suppress it.
-                // For empty classes like `class C {} // comment`, scan_end must
-                // be the closing `}` position, not node.end — otherwise a newline
-                // after `}` (before the next statement) causes us to incorrectly
-                // suppress the trailing comment that belongs to `}`.
-                let scan_end = class
-                    .members
-                    .nodes
-                    .first()
-                    .and_then(|&idx| self.arena.get(idx))
-                    .map_or_else(
-                        || {
-                            // Empty class: find the closing `}` to use as scan_end
-                            let be = brace_end as usize;
-                            if be <= end {
-                                bytes[be..end]
-                                    .iter()
-                                    .position(|&b| b == b'}')
-                                    .map_or(end, |p| be + p)
-                            } else {
-                                end
-                            }
-                        },
-                        |m| m.pos as usize,
-                    );
-                let brace_end_usize = brace_end as usize;
-                let scan_end_clamped = scan_end.min(end);
-                let has_newline = if brace_end_usize <= scan_end_clamped {
-                    bytes[brace_end_usize..scan_end_clamped]
-                        .iter()
-                        .any(|&b| b == b'\n' || b == b'\r')
-                } else {
-                    // Malformed source: first member pos precedes the opening
-                    // brace we found — skip the suppression heuristic.
-                    false
-                };
-                if has_newline {
-                    self.skip_trailing_same_line_comments(brace_end, node.end);
-                }
-            }
-        }
-        self.write_line();
-        self.increase_indent();
-        // When inside a comma expression wrapper (class expression with private fields
-        // or static fields), add one extra indent level for the class body to match tsc.
-        if class_expr_temp.is_some() {
-            self.increase_indent();
-        }
+        self.emit_class_es6_header_and_open_body(
+            node,
+            class,
+            assignment_prefix.is_some(),
+            static_super_base_alias.as_deref(),
+            class_expr_temp.is_some(),
+        );
 
         // Store auto-accessor inits for constructor emission.
         let prev_auto_accessor_inits = std::mem::take(&mut self.pending_auto_accessor_inits);
@@ -1870,167 +1596,20 @@ impl<'a> Printer<'a> {
         // (e.g., `[Symbol.iterator]: Type` → erased member, but `Symbol.iterator;` emitted).
         let mut computed_property_side_effects: Vec<NodeIndex> = Vec::new();
 
-        // Collect property initializers that need lowering
-        // (name, initializer_idx, init_end, leading_comments, trailing_comments)
-        // Comments are collected eagerly here so they're available even
-        // when the constructor appears before the property in source order.
-        let mut field_inits: Vec<crate::emitter::core::FieldInit> = Vec::new();
-        let mut static_field_inits: Vec<StaticFieldInit> = Vec::new();
-        let mut hoisted_native_private_members: FxHashSet<NodeIndex> = FxHashSet::default();
-        let mut hoisted_native_auto_accessor_members: FxHashSet<NodeIndex> = FxHashSet::default();
-        if needs_class_field_lowering {
-            let members = &class.members.nodes;
-            for (member_i, &member_idx) in members.iter().enumerate() {
-                if let Some(member_node) = self.arena.get(member_idx)
-                    && member_node.kind == syntax_kind_ext::PROPERTY_DECLARATION
-                    && let Some(prop) = self.arena.get_property_decl(member_node)
-                {
-                    // With useDefineForClassFields, fields without initializers
-                    // are still materialized at runtime as
-                    // `Object.defineProperty(this, "name", { value: void 0 })`.
-                    // Without that flag the typed-only declaration has no
-                    // runtime effect, so skip it.
-                    let no_initializer_node = prop.initializer.is_none();
-                    let materialize_no_init = self.no_init_property_is_runtime_materialized(prop);
-                    if !materialize_no_init
-                        && (no_initializer_node
-                            || !self.class_property_initializer_has_equals(member_node, prop))
-                    {
-                        continue;
-                    }
-                    if self
-                        .arena
-                        .has_modifier(&prop.modifiers, SyntaxKind::AbstractKeyword)
-                        || self
-                            .arena
-                            .has_modifier(&prop.modifiers, SyntaxKind::DeclareKeyword)
-                    {
-                        continue;
-                    }
-                    // Skip private fields when they're being lowered to WeakMap pattern.
-                    // They're handled separately via pending_private_field_constructor_inits.
-                    if !private_fields.is_empty() && is_private_identifier(self.arena, prop.name) {
-                        continue;
-                    }
-                    let is_private_name = is_private_identifier(self.arena, prop.name);
-                    let is_auto_accessor = self
-                        .arena
-                        .has_modifier(&prop.modifiers, SyntaxKind::AccessorKeyword);
-                    if !needs_private_field_lowering
-                        && is_private_name
-                        && (!hoist_native_instance_order_inits
-                            || self.has_effective_static_modifier_js(&prop.modifiers))
-                    {
-                        continue;
-                    }
-                    if is_auto_accessor
-                        && (!hoist_native_instance_order_inits
-                            || self.has_effective_static_modifier_js(&prop.modifiers))
-                    {
-                        continue;
-                    }
-                    // If the property has a computed name with a hoisted temp, use the temp
-                    // variable name. This takes priority over get_property_name_emit because
-                    // the temp captures the expression value at class-evaluation time.
-                    let mut name_emit = if let Some(name_node) = self.arena.get(prop.name)
-                        && name_node.kind == syntax_kind_ext::COMPUTED_PROPERTY_NAME
-                        && let Some(computed) = self.arena.get_computed_property(name_node)
-                        && let Some(temp) = self.computed_prop_temp_map.get(&computed.expression)
-                    {
-                        Some(PropertyNameEmit::Bracket(temp.clone()))
-                    } else {
-                        self.get_property_name_emit(prop.name)
-                    };
-                    if hoist_native_instance_order_inits && is_auto_accessor {
-                        if let Some((storage_name, _)) = auto_accessor_member_map.get(&member_idx) {
-                            name_emit = Some(PropertyNameEmit::Dot(format!("#{storage_name}")));
-                            hoisted_native_auto_accessor_members.insert(member_idx);
-                        }
-                    } else if hoist_native_instance_order_inits
-                        && is_private_name
-                        && let Some(private_name) = get_private_field_name(self.arena, prop.name)
-                    {
-                        name_emit = Some(PropertyNameEmit::Dot(private_name));
-                        hoisted_native_private_members.insert(member_idx);
-                    }
-                    let Some(name_emit) = name_emit else {
-                        continue;
-                    };
-
-                    // Pre-collect leading comments for this property declaration.
-                    // Use the actual token end of the previous member (not its
-                    // `end` field which can overshoot into the next member's trivia)
-                    // so the range doesn't invert.
-                    let leading_comments = if !self.ctx.options.remove_comments {
-                        let prev_end = if member_i > 0 {
-                            members
-                                .get(member_i - 1)
-                                .and_then(|&prev_idx| self.arena.get(prev_idx))
-                                .map_or(member_node.pos, |prev| {
-                                    self.find_token_end_before_trivia(prev.pos, prev.end)
-                                })
-                        } else {
-                            member_node.pos.saturating_sub(64)
-                        };
-                        self.collect_leading_comments_in_range(prev_end, member_node.pos)
-                    } else {
-                        Vec::new()
-                    };
-
-                    // Pre-collect trailing comments for this property declaration.
-                    let trailing_comments = if !self.ctx.options.remove_comments {
-                        let skip_end = members
-                            .get(member_i + 1)
-                            .and_then(|&next_idx| self.arena.get(next_idx))
-                            .map_or(member_node.end, |next| next.pos);
-                        let actual_end =
-                            self.find_token_end_before_trivia(member_node.pos, skip_end);
-                        self.collect_trailing_comments_in_range(actual_end)
-                    } else {
-                        Vec::new()
-                    };
-
-                    if self.has_effective_static_modifier_js(&prop.modifiers) {
-                        // `tsc` erases a no-init static field (see `static_field_erasure`).
-                        let erased = static_no_init_field_is_erased(
-                            prop.initializer.is_none(),
-                            self.ctx.options.use_define_for_class_fields,
-                        );
-                        if needs_static_block_lowering && !erased {
-                            static_field_inits.push((
-                                name_emit,
-                                prop.initializer,
-                                member_node.pos,
-                                Vec::new(), // leading_comments filled during class body emission
-                                Vec::new(), // trailing_comments filled during class body emission
-                            ));
-                        }
-                    } else {
-                        // Non-static field inits use String names for `this.name = val`,
-                        // `this["name"] = val`, or `this[0] = val`. Bracket names use
-                        // a `[` prefix to signal bracket notation at emit time.
-                        let ident_name = match &name_emit {
-                            PropertyNameEmit::Dot(s) => s.clone(),
-                            PropertyNameEmit::Bracket(s) | PropertyNameEmit::BracketNumeric(s) => {
-                                format!("[{s}]")
-                            }
-                        };
-                        let init_end = self
-                            .arena
-                            .get(prop.initializer)
-                            .map_or(member_node.end, |n| n.end);
-                        field_inits.push((
-                            ident_name,
-                            prop.initializer,
-                            init_end,
-                            leading_comments,
-                            trailing_comments,
-                            member_node.pos,
-                        ));
-                    }
-                }
-            }
-        }
+        let ClassFieldInitCollection {
+            mut field_inits,
+            mut static_field_inits,
+            hoisted_native_private_members,
+            hoisted_native_auto_accessor_members,
+        } = self.collect_class_es6_field_inits(
+            class,
+            needs_class_field_lowering,
+            needs_private_field_lowering,
+            hoist_native_instance_order_inits,
+            &auto_accessor_member_map,
+            needs_static_block_lowering,
+            &private_fields,
+        );
 
         // Check if class has an explicit constructor with a body.
         // A constructor without a body (e.g., broken syntax `constructor` with no
@@ -2179,12 +1758,6 @@ impl<'a> Printer<'a> {
             self.function_scope_depth -= 1;
         }
 
-        // When useDefineForClassFields is true AND target >= ES2022 (native class fields),
-        // emit parameter property field declarations (e.g. `foo;`) at the beginning of
-        // the class body. TSC emits these before any other class members.
-        // When target < ES2022, fields are lowered to the constructor body, so no
-        // class-body declaration is needed.
-        let mut emitted_any_member = false;
         let target_supports_native_fields =
             (self.ctx.options.target as u32) >= (ScriptTarget::ES2022 as u32);
         let target_supports_native_private_names =
@@ -2194,608 +1767,42 @@ impl<'a> Printer<'a> {
             && class.members.nodes.iter().any(|&member_idx| {
                 self.legacy_member_decorator_needs_private_name_scope(member_idx)
             });
-        emitted_any_member |= self.emit_synthetic_static_self_alias_block(node, assignment_alias);
-        if self.ctx.options.use_define_for_class_fields && target_supports_native_fields {
-            // Find the constructor and collect its parameter properties
-            for &member_idx in &class.members.nodes {
-                if let Some(member_node) = self.arena.get(member_idx)
-                    && member_node.kind == syntax_kind_ext::CONSTRUCTOR
-                    && let Some(ctor) = self.arena.get_constructor(member_node)
-                    && ctor.body.is_some()
-                {
-                    let param_props = self.collect_parameter_properties(&ctor.parameters.nodes);
-                    for name in &param_props {
-                        self.write(name);
-                        self.write(";");
-                        self.write_line();
-                        emitted_any_member = true;
-                    }
-                    break;
-                }
-            }
-        }
-        // Compute the class body's closing `}` position so the last member's
-        // trailing comment scan doesn't overshoot into comments belonging to
-        // the closing brace line (same pattern as namespace IIFE emitter).
-        let class_body_close_pos = self
-            .source_text
-            .map(|text| {
-                let end = std::cmp::min(node.end as usize, text.len());
-                let bytes = text.as_bytes();
-                let mut pos = end;
-                while pos > 0 {
-                    pos -= 1;
-                    if bytes[pos] == b'}' {
-                        return pos as u32;
-                    }
-                }
-                node.end
-            })
-            .unwrap_or(node.end);
 
-        let mut field_init_comment_idx = 0usize;
-        for (member_i, &member_idx) in class.members.nodes.iter().enumerate() {
-            // Skip private field declarations entirely when lowering to WeakMap pattern
-            if !private_fields.is_empty()
-                && let Some(member_node) = self.arena.get(member_idx)
-                && member_node.kind == syntax_kind_ext::PROPERTY_DECLARATION
-                && let Some(prop) = self.arena.get_property_decl(member_node)
-                && self
-                    .arena
-                    .get(prop.name)
-                    .is_some_and(|n| n.kind == SyntaxKind::PrivateIdentifier as u16)
-                && !private_duplicate_conflicts.is_conflicting(member_idx)
-            {
-                // Skip comments that belong to this erased member
-                if let Some(mn) = self.arena.get(member_idx) {
-                    let skip_end = class
-                        .members
-                        .nodes
-                        .get(member_i + 1)
-                        .and_then(|&next_idx| self.arena.get(next_idx))
-                        .map_or(mn.end, |next| next.pos);
-                    while self.comment_emit_idx < self.all_comments.len()
-                        && self.all_comments[self.comment_emit_idx].end <= skip_end
-                    {
-                        self.comment_emit_idx += 1;
-                    }
-                }
-                continue;
-            }
-            // Skip private methods and accessors that are extracted as standalone functions
-            if !self.private_members_to_skip.is_empty() {
-                let should_skip = self.private_members_to_skip.contains(&member_idx);
-                if should_skip {
-                    // When source has trailing `;` after private method/accessor
-                    // (e.g., `#foo() { };`), tsc preserves the semicolon.
-                    if let Some(mn) = self.arena.get(member_idx) {
-                        let has_trailing_semi = self.source_text.is_some_and(|text| {
-                            let start = mn.pos as usize;
-                            let end = std::cmp::min(mn.end as usize, text.len());
-                            if start >= end {
-                                return false;
-                            }
-                            let member_text = text[start..end].trim_end();
-                            if let Some(before_semi) = member_text.strip_suffix(';') {
-                                before_semi.trim_end().ends_with('}')
-                            } else {
-                                false
-                            }
-                        });
-                        if has_trailing_semi {
-                            if !self.writer.is_at_line_start() {
-                                self.write_line();
-                            }
-                            self.write(";");
-                            self.write_line();
-                            emitted_any_member = true;
-                        }
-                    }
-                    if let Some(mn) = self.arena.get(member_idx) {
-                        let skip_end = class
-                            .members
-                            .nodes
-                            .get(member_i + 1)
-                            .and_then(|&next_idx| self.arena.get(next_idx))
-                            .map_or(mn.end, |next| next.pos);
-                        while self.comment_emit_idx < self.all_comments.len()
-                            && self.all_comments[self.comment_emit_idx].end <= skip_end
-                        {
-                            self.comment_emit_idx += 1;
-                        }
-                    }
-                    continue;
-                }
-            }
-            // Skip property declarations that were lowered
-            if needs_class_field_lowering
-                && let Some(member_node) = self.arena.get(member_idx)
-                && member_node.kind == syntax_kind_ext::PROPERTY_DECLARATION
-                    && let Some(prop) = self.arena.get_property_decl(member_node)
-                    && !auto_accessor_member_map.contains_key(&member_idx)
-                    && prop.initializer.is_some()
-                    && !self
-                        .arena
-                        .has_modifier(&prop.modifiers, SyntaxKind::AbstractKeyword)
-                    // Auto-accessor properties (`accessor x = 1`) that are NOT being
-                    // lowered (e.g. at esnext target) must be preserved verbatim — they
-                    // are not regular field declarations.
-                    && !self
-                        .arena
-                        .has_modifier(&prop.modifiers, SyntaxKind::AccessorKeyword)
-                // Private fields (#name) are emitted verbatim at ES2022+ — they
-                // use native private field syntax and are unaffected by
-                // useDefineForClassFields.  Only skip them for lowering when the
-                // target actually requires WeakMap-based lowering (< ES2022).
-                && !(self.arena.get(prop.name).is_some_and(|n| {
-                    n.kind == SyntaxKind::PrivateIdentifier as u16
-                }) && (self.ctx.options.target as u32) >= (ScriptTarget::ES2022 as u32))
-                && !(self.arena.get(prop.name).is_some_and(|n| {
-                    n.kind == SyntaxKind::PrivateIdentifier as u16
-                }) && private_duplicate_conflicts.is_conflicting(member_idx))
-                // Static fields at ES2022+ are emitted inline as `static { this.f = v; }`
-                // blocks, not deferred to external assignments.
-                && (!self.has_effective_static_modifier_js(&prop.modifiers)
-                    || needs_static_block_lowering)
-            {
-                // For static properties, save leading and trailing comments before
-                // skipping so they can be emitted when the initialization is moved
-                // after the class body.
-                let is_static = self.has_effective_static_modifier_js(&prop.modifiers);
-                if is_static {
-                    let leading = self.collect_leading_comments(member_node.pos);
-                    if let Some(entry) = static_field_inits
-                        .iter_mut()
-                        .find(|e| e.2 == member_node.pos)
-                    {
-                        entry.3 = leading;
-                    }
-                }
-                if let Some(member_node) = self.arena.get(member_idx) {
-                    // Use a tighter bound for property declarations to avoid
-                    // consuming comments that belong to the next class member.
-                    // Property node.end can extend past newlines into the next
-                    // member's territory, so we bound by the next member's pos.
-                    let skip_end = class
-                        .members
-                        .nodes
-                        .get(member_i + 1)
-                        .and_then(|&next_idx| self.arena.get(next_idx))
-                        .map_or(member_node.end, |next| next.pos);
-                    // Find the actual end of the property's content
-                    let actual_end = self.find_token_end_before_trivia(member_node.pos, skip_end);
-                    // Find line end from actual_end
-                    let line_end = if let Some(text) = self.source_text {
-                        let bytes = text.as_bytes();
-                        let mut pos = actual_end as usize;
-                        while pos < bytes.len() && bytes[pos] != b'\n' && bytes[pos] != b'\r' {
-                            pos += 1;
-                        }
-                        pos as u32
-                    } else {
-                        actual_end
-                    };
-                    // Collect trailing comments on the same line for both static and
-                    // non-static fields. Static comments are stored on static_field_inits;
-                    // non-static comments are stored on field_inits for replay in the
-                    // constructor prologue.
-                    if let Some(text) = self.source_text {
-                        let mut trailing = Vec::new();
-                        let mut idx = self.comment_emit_idx;
-                        while idx < self.all_comments.len() {
-                            let c = &self.all_comments[idx];
-                            if c.pos >= actual_end
-                                && c.end <= line_end
-                                && let Ok(comment_text) =
-                                    crate::safe_slice::slice(text, c.pos as usize, c.end as usize)
-                            {
-                                trailing.push(comment_text.to_string());
-                            }
-                            if c.end > line_end {
-                                break;
-                            }
-                            idx += 1;
-                        }
-                        if is_static {
-                            if let Some(entry) = static_field_inits
-                                .iter_mut()
-                                .find(|e| e.2 == member_node.pos)
-                            {
-                                entry.4 = trailing;
-                            }
-                        } else if !trailing.is_empty() {
-                            if let Some(entry) = field_inits.get_mut(field_init_comment_idx) {
-                                entry.4 = trailing.clone();
-                            }
-                            // Also update pending_class_field_inits so existing constructors
-                            // that read from it during the member loop get the comments
-                            if let Some(entry) = self
-                                .pending_class_field_inits
-                                .get_mut(field_init_comment_idx)
-                            {
-                                entry.4 = trailing;
-                            }
-                        }
-                    }
-                    if !is_static {
-                        field_init_comment_idx += 1;
-                    }
-                    while self.comment_emit_idx < self.all_comments.len() {
-                        let c = &self.all_comments[self.comment_emit_idx];
-                        if c.end <= line_end {
-                            self.comment_emit_idx += 1;
-                        } else {
-                            break;
-                        }
-                    }
-                }
-                continue;
-            }
+        // When useDefineForClassFields is true AND target >= ES2022 (native class fields),
+        // emit parameter property field declarations (e.g. `foo;`) at the beginning of
+        // the class body. TSC emits these before any other class members.
+        // When target < ES2022, fields are lowered to the constructor body, so no
+        // class-body declaration is needed.
+        self.emit_class_es6_members(ClassEs6MemberEmit {
+            node,
+            class,
+            assignment_alias,
+            class_name: &class_name,
+            needs_class_field_lowering,
+            needs_static_block_lowering,
+            lower_auto_accessors_to_private_fields,
+            target_supports_native_fields,
+            target_supports_native_private_names,
+            has_legacy_private_name_member_decorators,
+            needs_computed_prop_hoisting,
+            private_fields: &private_fields,
+            private_duplicate_conflicts: &private_duplicate_conflicts,
+            auto_accessor_member_map: &auto_accessor_member_map,
+            auto_accessor_class_alias: auto_accessor_class_alias.as_deref(),
+            auto_accessor_computed_storage_key_member,
+            auto_accessor_instance_storage_inits_in_computed_key:
+                &auto_accessor_instance_storage_inits_in_computed_key,
+            hoisted_native_auto_accessor_members: &hoisted_native_auto_accessor_members,
+            hoisted_native_private_members: &hoisted_native_private_members,
+            field_inits: &mut field_inits,
+            static_field_inits: &mut static_field_inits,
+            deferred_static_blocks: &mut deferred_static_blocks,
+            computed_property_side_effects: &mut computed_property_side_effects,
+        });
 
-            // Skip static blocks that need lowering to IIFEs after the class
-            if needs_static_block_lowering
-                && let Some(member_node) = self.arena.get(member_idx)
-                && member_node.kind == syntax_kind_ext::CLASS_STATIC_BLOCK_DECLARATION
-            {
-                // Find the opening `{` of the static block to determine where
-                // inner (body) comments start. We skip leading comments but save
-                // the index of the first inner comment for replay during IIFE emission.
-                let brace_pos = if let Some(text) = self.source_text {
-                    let bytes = text.as_bytes();
-                    let start = member_node.pos as usize;
-                    let end = (member_node.end as usize).min(bytes.len());
-                    bytes[start..end]
-                        .iter()
-                        .position(|&b| b == b'{')
-                        .map(|off| (start + off + 1) as u32)
-                        .unwrap_or(member_node.end)
-                } else {
-                    member_node.end
-                };
-                // Skip comments preceding the block opening `{`
-                while self.comment_emit_idx < self.all_comments.len()
-                    && self.all_comments[self.comment_emit_idx].end <= brace_pos
-                {
-                    self.comment_emit_idx += 1;
-                }
-                // Save index pointing at the first inner comment (if any)
-                let inner_comment_idx = self.comment_emit_idx;
-                // Skip remaining inner comments so they don't leak as leading
-                // comments of subsequent class members
-                self.skip_comments_for_erased_node(member_node);
-                deferred_static_blocks.push((member_idx, inner_comment_idx));
-                continue;
-            }
-
-            // Check if this member is erased (no runtime representation)
-            if let Some(member_node) = self.arena.get(member_idx) {
-                let is_erased = match member_node.kind {
-                    // Bodyless methods are erased (abstract methods without body,
-                    // overload signatures). Abstract methods WITH a body (an error
-                    // in TS) are still emitted by tsc, so we must not erase them.
-                    k if k == syntax_kind_ext::METHOD_DECLARATION => {
-                        self.arena.get_method_decl(member_node).is_some_and(|m| {
-                            m.body.is_none()
-                                && !self.has_recovered_declaration_trailing_comma(member_node)
-                        })
-                    }
-                    // Abstract accessors without body are erased. Bodyless non-abstract
-                    // accessors (error case) are kept — tsc emits them as `{}`.
-                    // Abstract accessors WITH a body (error case) are also kept.
-                    k if k == syntax_kind_ext::GET_ACCESSOR
-                        || k == syntax_kind_ext::SET_ACCESSOR =>
-                    {
-                        self.arena.get_accessor(member_node).is_some_and(|a| {
-                            self.arena
-                                .has_modifier(&a.modifiers, SyntaxKind::AbstractKeyword)
-                                && a.body.is_none()
-                        })
-                    }
-                    k if k == syntax_kind_ext::PROPERTY_DECLARATION => {
-                        if let Some(p) = self.arena.get_property_decl(member_node) {
-                            // Abstract properties: erased
-                            if self
-                                .arena
-                                .has_modifier(&p.modifiers, SyntaxKind::AbstractKeyword)
-                            {
-                                true
-                            } else {
-                                // Type-only properties (no initializer, not private, not accessor): erased.
-                                // Native class-field emit keeps uninitialised properties only
-                                // when the target can represent class fields in the class body.
-                                if self.ctx.options.use_define_for_class_fields
-                                    && target_supports_native_fields
-                                {
-                                    false
-                                } else {
-                                    let is_private = self.arena.get(p.name).is_some_and(|n| {
-                                        n.kind == SyntaxKind::PrivateIdentifier as u16
-                                    });
-                                    let has_accessor = self
-                                        .arena
-                                        .has_modifier(&p.modifiers, SyntaxKind::AccessorKeyword);
-                                    p.initializer.is_none() && !is_private && !has_accessor
-                                }
-                            }
-                        } else {
-                            false
-                        }
-                    }
-                    // Bodyless constructor overloads are erased
-                    k if k == syntax_kind_ext::CONSTRUCTOR => self
-                        .arena
-                        .get_constructor(member_node)
-                        .is_some_and(|c| c.body.is_none()),
-                    // Index signatures are TypeScript-only
-                    k if k == syntax_kind_ext::INDEX_SIGNATURE => true,
-                    // Semicolon class elements are preserved in JS output (valid JS syntax)
-                    k if k == syntax_kind_ext::SEMICOLON_CLASS_ELEMENT => false,
-                    _ => false,
-                };
-                if is_erased {
-                    // When an erased property has a computed name whose expression
-                    // could have runtime side effects, tsc emits the expression as
-                    // a standalone statement after the class body.
-                    // e.g., `[Symbol.iterator]: Type` → `Symbol.iterator;`
-                    // Only expressions that might have observable effects are emitted:
-                    // property accesses, element accesses, calls, assignments, etc.
-                    // Simple identifiers and literals are NOT emitted (no side effects).
-                    // Skip this when computed property hoisting is active — the comma
-                    // expression already handles side effects.
-                    if !needs_computed_prop_hoisting
-                        && member_node.kind == syntax_kind_ext::PROPERTY_DECLARATION
-                        && let Some(p) = self.arena.get_property_decl(member_node)
-                        && let Some(name_node) = self.arena.get(p.name)
-                        && name_node.kind == syntax_kind_ext::COMPUTED_PROPERTY_NAME
-                        && let Some(computed) = self.arena.get_computed_property(name_node)
-                        && let Some(expr_node) = self.arena.get(computed.expression)
-                    {
-                        let k = expr_node.kind;
-                        let is_side_effect_free = k == SyntaxKind::Identifier as u16
-                            || k == SyntaxKind::StringLiteral as u16
-                            || k == SyntaxKind::NumericLiteral as u16
-                            || k == SyntaxKind::NoSubstitutionTemplateLiteral as u16
-                            || k == SyntaxKind::PrivateIdentifier as u16;
-                        if !is_side_effect_free {
-                            computed_property_side_effects.push(computed.expression);
-                        }
-                    }
-                    self.skip_comments_for_erased_node(member_node);
-                    continue;
-                }
-            }
-
-            // Emit leading comments before this member
-            if let Some(member_node) = self.arena.get(member_idx) {
-                let comments_end_with_line_break =
-                    self.pending_comments_before_pos_end_with_line_break(member_node.pos);
-                self.emit_comments_before_pos(member_node.pos);
-                if comments_end_with_line_break && !self.writer.is_at_line_start() {
-                    self.write_line();
-                }
-            }
-
-            let before_len = self.writer.len();
-            let auto_accessor = auto_accessor_member_map.get(&member_idx).cloned();
-            if let Some(member_node) = self.arena.get(member_idx) {
-                let property_end = if auto_accessor.is_some() {
-                    let upper = class
-                        .members
-                        .nodes
-                        .get(member_i + 1)
-                        .and_then(|&next_idx| self.arena.get(next_idx))
-                        .map(|n| n.pos)
-                        .unwrap_or(member_node.end);
-                    Some(self.find_token_end_before_trivia(member_node.pos, upper))
-                } else {
-                    None
-                };
-
-                if let Some((storage_name, is_static)) = auto_accessor {
-                    let computed_storage_inits =
-                        if Some(member_idx) == auto_accessor_computed_storage_key_member {
-                            auto_accessor_instance_storage_inits_in_computed_key.as_slice()
-                        } else {
-                            &[]
-                        };
-                    self.emit_auto_accessor_methods(
-                        member_node,
-                        &storage_name,
-                        is_static,
-                        AutoAccessorEmitOptions {
-                            static_accessor_alias: auto_accessor_class_alias.as_deref(),
-                            lower_to_private_fields: lower_auto_accessors_to_private_fields,
-                            class_name: &class_name,
-                            property_end: property_end.unwrap_or(member_node.end),
-                            omit_storage_initializer: hoisted_native_auto_accessor_members
-                                .contains(&member_idx),
-                            computed_storage_inits,
-                        },
-                    );
-                } else if hoisted_native_private_members.contains(&member_idx) {
-                    if let Some(prop) = self.arena.get_property_decl(member_node) {
-                        self.emit_class_member_modifiers_js(&prop.modifiers);
-                        if let Some(private_name) = get_private_field_name(self.arena, prop.name) {
-                            self.write(&private_name);
-                        }
-                        self.write_semicolon();
-                    }
-                } else {
-                    self.class_member_emit_depth = self.class_member_emit_depth.saturating_add(1);
-                    self.emit(member_idx);
-                    self.class_member_emit_depth = self.class_member_emit_depth.saturating_sub(1);
-                }
-            }
-            let mut emit_standalone_class_semicolon = false;
-            if let Some(member_node) = self.arena.get(member_idx)
-                && (member_node.kind == syntax_kind_ext::GET_ACCESSOR
-                    || member_node.kind == syntax_kind_ext::SET_ACCESSOR
-                    || member_node.kind == syntax_kind_ext::METHOD_DECLARATION)
-            {
-                let next_is_semicolon_member = class
-                    .members
-                    .nodes
-                    .get(member_i + 1)
-                    .and_then(|&idx| self.arena.get(idx))
-                    .is_some_and(|n| n.kind == syntax_kind_ext::SEMICOLON_CLASS_ELEMENT);
-
-                // Check if the member has a body (method/accessor with `{}`).
-                let member_has_body_for_semi = match member_node.kind {
-                    k if k == syntax_kind_ext::METHOD_DECLARATION => self
-                        .arena
-                        .get_method_decl(member_node)
-                        .is_some_and(|m| m.body.is_some()),
-                    k if k == syntax_kind_ext::GET_ACCESSOR
-                        || k == syntax_kind_ext::SET_ACCESSOR =>
-                    {
-                        self.arena
-                            .get_accessor(member_node)
-                            .is_some_and(|a| a.body.is_some())
-                    }
-                    _ => false,
-                };
-                if !next_is_semicolon_member {
-                    let has_source_semicolon = self.source_text.is_some_and(|text| {
-                        let member_end = std::cmp::min(member_node.end as usize, text.len());
-                        // For members WITHOUT bodies, check the gap after the member.
-                        if !member_has_body_for_semi {
-                            let gap_end = class
-                                .members
-                                .nodes
-                                .get(member_i + 1)
-                                .and_then(|&idx| self.arena.get(idx))
-                                .map_or_else(
-                                    || {
-                                        let search_end =
-                                            std::cmp::min(node.end as usize, text.len());
-                                        text[member_end..search_end]
-                                            .rfind('}')
-                                            .map_or(search_end, |pos| member_end + pos)
-                                    },
-                                    |n| n.pos as usize,
-                                );
-                            let gap_end = std::cmp::min(gap_end, text.len());
-                            if member_end < gap_end && text[member_end..gap_end].contains(';') {
-                                return true;
-                            }
-                        }
-                        // For members WITH bodies, the parser may absorb trailing `;`
-                        // into the member span (e.g., `get x() { ... };`).
-                        // Check if the member source ends with `} ;` pattern.
-                        if member_has_body_for_semi && member_end >= 2 {
-                            let tail = &text[member_node.pos as usize..member_end];
-                            let trimmed = tail.trim_end();
-                            if let Some(before_semi) = trimmed.strip_suffix(';')
-                                && before_semi.trim_end().ends_with('}')
-                            {
-                                return true;
-                            }
-                        }
-                        false
-                    });
-                    emit_standalone_class_semicolon = has_source_semicolon;
-                }
-
-                // Some parser recoveries include the semicolon in member.end without
-                // creating a separate SEMICOLON_CLASS_ELEMENT; preserve it from source.
-                // Only check this for methods/accessors that DON'T have a body (i.e.,
-                // abstract methods or overload signatures like `foo(): void;`).
-                if !member_has_body_for_semi
-                    && self.source_text.is_some_and(|text| {
-                        let start = std::cmp::min(member_node.pos as usize, text.len());
-                        let end = std::cmp::min(member_node.end as usize, text.len());
-                        if start >= end {
-                            return false;
-                        }
-                        let member_text = text[start..end].trim_end();
-                        member_text.ends_with(';')
-                    })
-                {
-                    emit_standalone_class_semicolon = true;
-                }
-            }
-            if self.writer.len() == before_len
-                && let (Some(member_node), Some(text)) =
-                    (self.arena.get(member_idx), self.source_text)
-            {
-                let start = std::cmp::min(member_node.pos as usize, text.len());
-                let end = std::cmp::min(member_node.end as usize, text.len());
-                if start < end {
-                    let raw = &text[start..end];
-                    let compact: String = raw.chars().filter(|c| !c.is_whitespace()).collect();
-                    if compact.starts_with("*(){") {
-                        self.write("*() { }");
-                    }
-                }
-            }
-            // Only add newline if something was actually emitted
-            if self.writer.len() > before_len && !self.writer.is_at_line_start() {
-                emitted_any_member = true;
-                // Emit trailing comments on the same line as the member.
-                // For property declarations, member_node.end can include the leading trivia
-                // of the next member (because the parser records token_end() = scanner.pos
-                // which is after the lookahead token). Use the AST initializer/name end
-                // to get the true end of the property's last token.
-                if let Some(member_node) = self.arena.get(member_idx) {
-                    // Use the next member's pos as upper bound to avoid scanning
-                    // past the current member into the next member's trivia.
-                    // For the last member, use the class body's closing `}` position
-                    // so we don't steal comments that belong on the closing brace line.
-                    let next_member_pos = class
-                        .members
-                        .nodes
-                        .get(member_i + 1)
-                        .and_then(|&next_idx| self.arena.get(next_idx))
-                        .map(|n| n.pos);
-                    let upper = next_member_pos.unwrap_or(member_node.end);
-                    let token_end = self.find_token_end_before_trivia(member_node.pos, upper);
-                    // For the last member, cap trailing comment scan at the class
-                    // body's closing `}` to avoid stealing comments that belong
-                    // on the closing brace line.
-                    if next_member_pos.is_none() {
-                        self.emit_trailing_comments_before(token_end, class_body_close_pos);
-                    } else {
-                        self.emit_trailing_comments(token_end);
-                    }
-                }
-                self.write_line();
-                if emit_standalone_class_semicolon {
-                    self.write(";");
-                    self.write_line();
-                }
-                if target_supports_native_private_names
-                    && has_legacy_private_name_member_decorators
-                    && self.legacy_member_decorator_needs_private_name_scope(member_idx)
-                {
-                    self.write("static {");
-                    self.write_line();
-                    self.increase_indent();
-                    self.emit_legacy_member_decorator_calls_requiring_private_name_scope(
-                        &class_name,
-                        &[member_idx],
-                    );
-                    self.decrease_indent();
-                    self.write("}");
-                    self.write_line();
-                }
-            }
-        }
         self.scoped_class_expression_self_alias_ancestors
             .truncate(scoped_class_expression_self_alias_ancestor_len);
         self.scoped_class_expression_self_alias = prev_scoped_class_expression_self_alias;
-
-        if !emitted_any_member && let Some(text) = self.source_text {
-            let start = std::cmp::min(node.pos as usize, text.len());
-            let end = std::cmp::min(node.end as usize, text.len());
-            if start < end {
-                let raw = &text[start..end];
-                let compact: String = raw.chars().filter(|c| !c.is_whitespace()).collect();
-                if compact.contains("*(){}") {
-                    self.write("*() { }");
-                    self.write_line();
-                }
-            }
-        }
 
         let computed_side_effects_emitted_in_static_block =
             !computed_property_side_effects.is_empty() && class_expr_temp.is_none();
@@ -2878,980 +1885,43 @@ impl<'a> Printer<'a> {
             self.write(";");
         }
 
-        // Emit computed property name hoisting comma expression or standalone side effects.
-        if !computed_prop_entries.is_empty() {
-            if class_expr_temp.is_some() {
-                for (entry_idx, (temp_name, expr_idx, _)) in
-                    computed_prop_entries.iter().enumerate()
-                {
-                    if computed_prop_entries_consumed_by_member_name.contains(&entry_idx) {
-                        continue;
-                    }
-                    self.write(",");
-                    self.write_line();
-                    self.increase_indent();
-                    if let Some(temp) = temp_name {
-                        self.write(temp);
-                        self.write(" = ");
-                    }
-                    self.emit_expression(*expr_idx);
-                    self.decrease_indent();
-                }
-            } else if computed_prop_entries
-                .iter()
-                .enumerate()
-                .any(|(entry_idx, _)| {
-                    !computed_prop_entries_consumed_by_member_name.contains(&entry_idx)
-                })
-            {
-                // Emit as a single comma expression: `_a = expr1, sideEffect, _b = expr2;`
-                self.write_line();
-                let mut emitted_entry = false;
-                for (entry_idx, (temp_name, expr_idx, _)) in
-                    computed_prop_entries.iter().enumerate()
-                {
-                    if computed_prop_entries_consumed_by_member_name.contains(&entry_idx) {
-                        continue;
-                    }
-                    if emitted_entry {
-                        self.write(", ");
-                    }
-                    emitted_entry = true;
-                    if let Some(temp) = temp_name {
-                        self.write(temp);
-                        self.write(" = ");
-                    }
-                    self.emit_expression(*expr_idx);
-                }
-                if emitted_entry {
-                    self.write(";");
-                }
-            }
-            if needs_computed_prop_comma_expr
-                && !needs_static_comma_expr
-                && !needs_private_comma_expr
-                && let Some(temp) = class_expr_temp.as_ref()
-            {
-                self.write(",");
-                self.write_line();
-                self.increase_indent();
-                self.write(temp);
-                if class_expr_comma_needs_parens {
-                    self.write(")");
-                }
-                self.decrease_indent();
-            }
-        } else if !computed_side_effects_emitted_in_static_block {
-            // Emit computed property name side-effect statements for erased members
-            // (when hoisting is not active, e.g., ES2022+ targets).
-            // e.g., `[Symbol.iterator]: Type` → `Symbol.iterator;`
-            for expr_idx in &computed_property_side_effects {
-                if class_expr_temp.is_some() {
-                    self.write(",");
-                    self.write_line();
-                    self.increase_indent();
-                    self.emit_expression(*expr_idx);
-                    self.decrease_indent();
-                } else {
-                    self.write_line();
-                    self.emit_expression(*expr_idx);
-                    self.write(";");
-                }
-            }
-        }
-
-        if let Some(recovery_name) = self.class_var_function_recovery_name(node) {
-            self.write_line();
-            self.write("var ");
-            self.write(&recovery_name);
-            self.write(";");
-            self.write_line();
-            self.write("() => { };");
-        }
-
-        // Emit static field initializers after class body. Class expressions use
-        // a comma expression; class declarations use separate statements.
-        //
-        // For a class *declaration* lowered to the WeakMap pattern, tsc always
-        // emits the private-member init statements before the static field
-        // assignment statements: a static initializer can instantiate the class,
-        // whose constructor populates the WeakMaps, so the storage must exist
-        // first. This holds even without a class/private self-reference, so the
-        // gate fires whenever private lowering and static elements coexist on a
-        // declaration, not only for the self-referential subset.
-        let emit_private_inits_before_static_elements = !needs_private_comma_expr
-            && has_any_private_lowering
-            && (!static_field_inits.is_empty() || !deferred_static_blocks.is_empty());
-        let mut emitted_private_auto_accessors_pre_static = false;
-        if emit_private_inits_before_static_elements {
-            let static_private_inits = std::mem::take(&mut self.pending_static_private_inits);
-            let private_class_alias_pair = self.pending_private_class_alias.take();
-            let instances_ws = self.pending_instances_weakset_add.take();
-            let method_defs = std::mem::take(&mut self.pending_private_method_defs);
-            let accessor_defs = std::mem::take(&mut self.pending_private_accessor_defs);
-            // Consume the pending WeakMap inits here so the later post-class
-            // emission path (which re-checks `pending_weakmap_inits`) does not
-            // emit the same `_X = new WeakMap()` lines a second time.
-            let weakmap_inits = std::mem::take(&mut self.pending_weakmap_inits);
-            let private_auto_instance_storage_inits: Vec<String> = private_auto_accessors
-                .iter()
-                .filter(|a| !a.is_static)
-                .map(|a| format!("{} = new WeakMap()", a.storage_name))
-                .collect();
-            let has_pre_static_private_inits = private_class_alias_pair.is_some()
-                || !weakmap_inits.is_empty()
-                || instances_ws.is_some()
-                || !method_defs.is_empty()
-                || !accessor_defs.is_empty()
-                || !private_auto_accessors.is_empty()
-                || !private_auto_instance_storage_inits.is_empty()
-                || !static_private_inits.is_empty();
-
-            if has_pre_static_private_inits {
-                self.write_line();
-                let mut first = true;
-                if let Some((ref alias, ref cls_name)) = private_class_alias_pair {
-                    self.write(alias);
-                    self.write(" = ");
-                    self.write(cls_name);
-                    first = false;
-                }
-                for init in &weakmap_inits {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.write(init);
-                    first = false;
-                }
-                if let Some(ref ws_name) = instances_ws {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.write(ws_name);
-                    self.write(" = new WeakSet()");
-                    first = false;
-                }
-                for init in &private_auto_instance_storage_inits {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.write(init);
-                    first = false;
-                }
-                for def in &method_defs {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.emit_private_method_function_def(
-                        def,
-                        private_member_def_needs_class_alias,
-                        class_value_alias.as_deref(),
-                        &class_name,
-                    );
-                    first = false;
-                }
-                for def in &accessor_defs {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.emit_private_accessor_function_def(
-                        def,
-                        private_member_def_needs_class_alias,
-                        class_value_alias.as_deref(),
-                        &class_name,
-                    );
-                    first = false;
-                }
-                for accessor in &private_auto_accessors {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.emit_private_auto_accessor_function_def(
-                        &accessor.get_var_name,
-                        &accessor.storage_name,
-                        accessor.is_static,
-                        true,
-                        private_class_alias_pair
-                            .as_ref()
-                            .map(|(alias, _)| alias.as_str())
-                            .or(class_value_alias.as_deref()),
-                    );
-                    self.write(", ");
-                    self.emit_private_auto_accessor_function_def(
-                        &accessor.set_var_name,
-                        &accessor.storage_name,
-                        accessor.is_static,
-                        false,
-                        private_class_alias_pair
-                            .as_ref()
-                            .map(|(alias, _)| alias.as_str())
-                            .or(class_value_alias.as_deref()),
-                    );
-                    first = false;
-                }
-                if !private_auto_accessors.is_empty() {
-                    emitted_private_auto_accessors_pre_static = true;
-                }
-                self.write(";");
-                for init in &static_private_inits {
-                    self.write_line();
-                    self.emit_static_private_init(init, &class_name, true);
-                }
-            }
-        }
-        // Private helper/state initialization can be part of a class-expression
-        // comma list. Gather it before static element scheduling so lowered
-        // private state can be emitted before static field/block work observes it.
-        let weakmap_inits = std::mem::take(&mut self.pending_weakmap_inits);
-        let has_weakmap_inits = !weakmap_inits.is_empty();
-        let static_private_inits = std::mem::take(&mut self.pending_static_private_inits);
-        let private_class_alias_pair = self.pending_private_class_alias.take();
-        let instances_ws = self.pending_instances_weakset_add.clone();
-        let method_defs: Vec<PrivateMethodDef> =
-            std::mem::take(&mut self.pending_private_method_defs);
-        let accessor_defs: Vec<PrivateAccessorDef> =
-            std::mem::take(&mut self.pending_private_accessor_defs);
-        let private_auto_instance_storage_inits: Vec<String> = private_auto_accessors
-            .iter()
-            .filter(|_| !emitted_private_auto_accessors_pre_static)
-            .filter(|a| !a.is_static)
-            .map(|a| format!("{} = new WeakMap()", a.storage_name))
-            .collect();
-        let has_post_class_inits = private_class_alias_pair.is_some()
-            || has_weakmap_inits
-            || instances_ws.is_some()
-            || !method_defs.is_empty()
-            || !accessor_defs.is_empty()
-            || !private_auto_instance_storage_inits.is_empty();
-
-        let class_expr_static_comma_had_scheduled_elements =
-            !static_field_inits.is_empty() || !deferred_static_blocks.is_empty();
-        let mut emitted_private_comma_items_before_static_items = false;
-        if !static_field_inits.is_empty()
-            && let Some(temp) = class_expr_static_temp.as_ref()
-        {
-            // Class expression comma-expression: `(_a = class C {}, _a.a = 1, _a)`
-            // The `(_a = ` prefix was already emitted before the `class` keyword.
-            //
-            // Static field initializers and (when not deferred) static blocks
-            // must be interleaved by source position so that observable
-            // evaluation order matches the source — e.g.
-            // `static a = 1; static { console.log(this.a); } static b = 2;`
-            // must emit the static block AFTER `_a.a = 1` and BEFORE `_a.b = 2`.
-            // Devin review: <https://github.com/mohsen1/tsz/pull/2279#discussion_r3176494185>
-            //
-            // We build a single position-keyed list. `field` items reuse the
-            // owned `StaticFieldInit` entries; `block` items consume the
-            // `(NodeIndex, usize)` deferred entries. When static blocks are
-            // deferred (via `--useDefineForClassFields` lowering deferral),
-            // they're emitted in their existing trailing batch instead.
-            let interleave_blocks = !self.defer_class_static_blocks;
-            enum CommaItem {
-                SetFunctionName(String),
-                Field(StaticFieldInit),
-                Block(NodeIndex, usize),
-            }
-            let owned_field_inits = std::mem::take(&mut static_field_inits);
-            let mut comma_items: Vec<(u32, CommaItem)> = Vec::new();
-            if let Some(name) = class_expr_set_function_name.as_ref() {
-                comma_items.push((node.pos, CommaItem::SetFunctionName(name.clone())));
-            }
-            comma_items.extend(
-                owned_field_inits
-                    .into_iter()
-                    .map(|init| (init.2, CommaItem::Field(init))),
-            );
-            if interleave_blocks {
-                let blocks = std::mem::take(&mut deferred_static_blocks);
-                for (block_idx, comment_idx) in blocks {
-                    let pos = self.arena.get(block_idx).map_or(u32::MAX, |node| node.pos);
-                    comma_items.push((pos, CommaItem::Block(block_idx, comment_idx)));
-                }
-            }
-            comma_items.sort_by_key(|(pos, _)| *pos);
-
-            if needs_private_comma_expr && has_post_class_inits {
-                emitted_private_comma_items_before_static_items = true;
-                self.emit_private_comma_items(PrivateCommaItems {
-                    weakmap_inits: &weakmap_inits,
-                    instances_ws: instances_ws.as_deref(),
-                    private_auto_instance_storage_inits: &private_auto_instance_storage_inits,
-                    method_defs: &method_defs,
-                    accessor_defs: &accessor_defs,
-                    private_member_def_needs_class_alias,
-                    class_value_alias: class_value_alias.as_deref(),
-                    class_name: &class_name,
-                    emitted_private_auto_accessors_pre_static,
-                    private_auto_accessors: &private_auto_accessors,
-                    private_class_alias_pair: private_class_alias_pair.as_ref(),
-                    set_function_name: class_expr_set_function_name
-                        .as_deref()
-                        .map(|name| (temp.as_str(), name)),
-                    static_private_inits: &static_private_inits,
-                });
-            }
-
-            for (_pos, item) in comma_items {
-                match item {
-                    CommaItem::SetFunctionName(name) => {
-                        if emitted_private_comma_items_before_static_items {
-                            continue;
-                        }
-                        self.emit_class_expr_set_function_name_comma_item(temp, &name);
-                    }
-                    CommaItem::Field((
-                        name_emit,
-                        init_idx,
-                        _member_pos,
-                        _leading_comments,
-                        _trailing_comments,
-                    )) => {
-                        self.write(",");
-                        self.write_line();
-                        self.increase_indent();
-                        if self.ctx.options.use_define_for_class_fields {
-                            let define_name = match &name_emit {
-                                PropertyNameEmit::Dot(s) => format!("\"{s}\""),
-                                PropertyNameEmit::Bracket(s)
-                                | PropertyNameEmit::BracketNumeric(s) => s.clone(),
-                            };
-                            self.write("Object.defineProperty(");
-                            self.write(temp);
-                            self.write(", ");
-                            self.write(&define_name);
-                            self.write(", {");
-                            self.write_line();
-                            self.increase_indent();
-                            self.write("enumerable: true,");
-                            self.write_line();
-                            self.write("configurable: true,");
-                            self.write_line();
-                            self.write("writable: true,");
-                            self.write_line();
-                            self.write("value: ");
-                            // Emit the initializer, then substitute class name with temp var
-                            let before = self.writer.len();
-                            self.with_scoped_static_initializer_context_cleared(|this| {
-                                this.emit_expression(init_idx);
-                            });
-                            let after = self.writer.len();
-                            if !class_name.is_empty() && class_name != *temp {
-                                let full = self.writer.get_output().to_string();
-                                let segment = &full[before..after];
-                                let replaced = replace_identifier(segment, &class_name, temp);
-                                if replaced != segment {
-                                    self.writer.truncate(before);
-                                    self.write(&replaced);
-                                }
-                            }
-                            self.write_line();
-                            self.decrease_indent();
-                            self.write("})");
-                        } else {
-                            self.write(temp);
-                            match &name_emit {
-                                PropertyNameEmit::Dot(name) => {
-                                    self.write(".");
-                                    self.write(name);
-                                }
-                                PropertyNameEmit::Bracket(name)
-                                | PropertyNameEmit::BracketNumeric(name) => {
-                                    self.write("[");
-                                    self.write(name);
-                                    self.write("]");
-                                }
-                            }
-                            self.write(" = ");
-                            let before = self.writer.len();
-                            self.with_scoped_static_initializer_context_cleared(|this| {
-                                this.emit_expression(init_idx);
-                            });
-                            let after = self.writer.len();
-                            if !class_name.is_empty() && class_name != *temp {
-                                let full = self.writer.get_output().to_string();
-                                let segment = &full[before..after];
-                                let replaced = replace_identifier(segment, &class_name, temp);
-                                if replaced != segment {
-                                    self.writer.truncate(before);
-                                    self.write(&replaced);
-                                }
-                            }
-                        }
-                        self.decrease_indent();
-                    }
-                    CommaItem::Block(block_idx, comment_idx) => {
-                        self.write(",");
-                        self.write_line();
-                        self.increase_indent();
-                        let prev_self_alias = self.scoped_class_expression_self_alias.clone();
-                        if class_name_is_real && !class_name.is_empty() && class_name != *temp {
-                            self.scoped_class_expression_self_alias = Some((
-                                Arc::<str>::from(class_name.as_str()),
-                                Arc::<str>::from(temp.as_str()),
-                            ));
-                        }
-                        self.emit_static_block_iife_expression(block_idx, comment_idx);
-                        self.scoped_class_expression_self_alias = prev_self_alias;
-                        self.decrease_indent();
-                    }
-                }
-            }
-            self.write(",");
-            self.write_line();
-            self.increase_indent();
-            self.write(temp);
-            if class_expr_comma_needs_parens {
-                self.write(")");
-            }
-            self.decrease_indent();
-            if assignment_prefix.is_some() {
-                self.write(";");
-            }
-        } else if !static_field_inits.is_empty() && !class_name.is_empty() {
-            self.write_line();
-            if let Some(temp) = default_export_set_function_name_temp.as_ref() {
-                self.write_helper("__setFunctionName");
-                self.write("(");
-                self.write(temp);
-                self.write(", \"default\");");
-                self.write_line();
-            }
-            // If lowered static elements need a stable class value, emit
-            // `_a = ClassName;` so `this` and class-name references can use it.
-            if !emit_private_inits_before_static_elements
-                && let Some(ref alias) = static_initializer_class_alias
-            {
-                self.write(alias);
-                self.write(" = ");
-                self.write(&class_name);
-                self.write(";");
-                self.write_line();
-            }
-            let mut next_static_block = 0usize;
-            for (name_emit, init_idx, _member_pos, leading_comments, trailing_comments) in
-                &static_field_inits
-            {
-                if !self.defer_class_static_blocks {
-                    while next_static_block < deferred_static_blocks.len() {
-                        let (block_idx, comment_idx) = deferred_static_blocks[next_static_block];
-                        let block_pos = self.arena.get(block_idx).map_or(u32::MAX, |node| node.pos);
-                        if block_pos >= *_member_pos {
-                            break;
-                        }
-                        let prev_this_alias = self.scoped_static_this_alias.clone();
-                        let prev_super_alias = self.scoped_static_super_base_alias.clone();
-                        self.scoped_static_this_alias =
-                            static_initializer_this_binding.map(std::sync::Arc::from);
-                        self.scoped_static_super_base_alias =
-                            static_initializer_super_base.map(std::sync::Arc::from);
-                        let prev_self_alias = self.scoped_class_expression_self_alias.clone();
-                        if let Some(alias) = static_initializer_class_alias.as_ref() {
-                            self.scoped_class_expression_self_alias = Some((
-                                Arc::<str>::from(class_name.as_str()),
-                                Arc::<str>::from(alias.as_str()),
-                            ));
-                        }
-                        self.emit_static_block_iife_expression(block_idx, comment_idx);
-                        self.scoped_class_expression_self_alias = prev_self_alias;
-                        self.scoped_static_this_alias = prev_this_alias;
-                        self.scoped_static_super_base_alias = prev_super_alias;
-                        self.write(";");
-                        self.write_line();
-                        next_static_block += 1;
-                    }
-                }
-
-                // Emit saved leading comments from the original static property declaration
-                for (comment_text, source_pos) in leading_comments {
-                    self.write_comment_with_reindent(comment_text, Some(*source_pos));
-                    self.write_line();
-                }
-                if self.ctx.options.use_define_for_class_fields {
-                    let define_name = match name_emit {
-                        PropertyNameEmit::Dot(s) => format!("\"{s}\""),
-                        PropertyNameEmit::Bracket(s) | PropertyNameEmit::BracketNumeric(s) => {
-                            s.clone()
-                        }
-                    };
-                    self.write("Object.defineProperty(");
-                    self.write(&class_name);
-                    self.write(", ");
-                    self.write(&define_name);
-                    self.write(", {");
-                    self.write_line();
-                    self.increase_indent();
-                    self.write("enumerable: true,");
-                    self.write_line();
-                    self.write("configurable: true,");
-                    self.write_line();
-                    self.write("writable: true,");
-                    self.write_line();
-                    self.write("value: ");
-                    let before = self.writer.len();
-                    self.emit_static_field_initializer_with_inner_comments(
-                        *init_idx,
-                        static_initializer_this_binding,
-                        static_initializer_super_base,
-                        externalized_static_initializer_uses_undefined_receiver,
-                    );
-                    let after = self.writer.len();
-                    if let Some(alias) =
-                        static_initializer_self_alias.or(static_initializer_class_alias.as_deref())
-                        && !class_name.is_empty()
-                        && class_name != alias
-                    {
-                        let full = self.writer.get_output().to_string();
-                        let segment = &full[before..after];
-                        let replaced = replace_identifier(segment, &class_name, alias);
-                        if replaced != segment {
-                            self.writer.truncate(before);
-                            self.write(&replaced);
-                        }
-                    }
-                    self.write_line();
-                    self.decrease_indent();
-                    self.write("});");
-                } else {
-                    self.write(&class_name);
-                    match name_emit {
-                        PropertyNameEmit::Dot(name) => {
-                            self.write(".");
-                            self.write(name);
-                        }
-                        PropertyNameEmit::Bracket(name)
-                        | PropertyNameEmit::BracketNumeric(name) => {
-                            self.write("[");
-                            self.write(name);
-                            self.write("]");
-                        }
-                    }
-                    self.write(" = ");
-                    let before = self.writer.len();
-                    self.emit_static_field_initializer_with_inner_comments(
-                        *init_idx,
-                        static_initializer_this_binding,
-                        static_initializer_super_base,
-                        externalized_static_initializer_uses_undefined_receiver,
-                    );
-                    let after = self.writer.len();
-                    if let Some(alias) =
-                        static_initializer_self_alias.or(static_initializer_class_alias.as_deref())
-                        && !class_name.is_empty()
-                        && class_name != alias
-                    {
-                        let full = self.writer.get_output().to_string();
-                        let segment = &full[before..after];
-                        let replaced = replace_identifier(segment, &class_name, alias);
-                        if replaced != segment {
-                            self.writer.truncate(before);
-                            self.write(&replaced);
-                        }
-                    }
-                    self.write(";");
-                }
-                // Emit saved trailing comments (e.g. `// ok` from
-                // `static intance = new C3(); // ok`)
-                for comment_text in trailing_comments {
-                    self.write_space();
-                    self.write_comment(comment_text);
-                }
-                self.write_line();
-            }
-            if !self.defer_class_static_blocks {
-                while next_static_block < deferred_static_blocks.len() {
-                    let (block_idx, comment_idx) = deferred_static_blocks[next_static_block];
-                    let prev_this_alias = self.scoped_static_this_alias.clone();
-                    let prev_super_alias = self.scoped_static_super_base_alias.clone();
-                    self.scoped_static_this_alias =
-                        static_initializer_this_binding.map(std::sync::Arc::from);
-                    self.scoped_static_super_base_alias =
-                        static_initializer_super_base.map(std::sync::Arc::from);
-                    let prev_self_alias = self.scoped_class_expression_self_alias.clone();
-                    if let Some(alias) = static_initializer_class_alias.as_ref() {
-                        self.scoped_class_expression_self_alias = Some((
-                            Arc::<str>::from(class_name.as_str()),
-                            Arc::<str>::from(alias.as_str()),
-                        ));
-                    }
-                    self.emit_static_block_iife_expression(block_idx, comment_idx);
-                    self.scoped_class_expression_self_alias = prev_self_alias;
-                    self.scoped_static_this_alias = prev_this_alias;
-                    self.scoped_static_super_base_alias = prev_super_alias;
-                    self.write(";");
-                    self.write_line();
-                    next_static_block += 1;
-                }
-                if next_static_block > 0 {
-                    deferred_static_blocks.clear();
-                }
-            }
-        }
-
-        let class_expr_static_comma_has_no_scheduled_elements =
-            class_expr_static_temp.is_some() && !class_expr_static_comma_had_scheduled_elements;
-        if class_expr_static_comma_has_no_scheduled_elements
-            && !needs_private_comma_expr
-            && let Some(temp) = class_expr_static_temp.as_ref()
-        {
-            if let Some(name) = class_expr_set_function_name.as_ref() {
-                self.emit_class_expr_set_function_name_comma_item(temp, name);
-            }
-            self.write(",");
-            self.write_line();
-            self.increase_indent();
-            self.write(temp);
-            if class_expr_comma_needs_parens {
-                self.write(")");
-            }
-            self.decrease_indent();
-            if assignment_prefix.is_some() {
-                self.write(";");
-            }
-        }
-
-        // Emit auto-accessor WeakMap initializations after class body:
-        // var _Class_prop_accessor_storage;
-        // ...
-        // _Class_prop_accessor_storage = new WeakMap();
-        if lower_auto_accessors_to_weakmap
-            && ((emit_auto_accessor_instance_inits_after_class
-                && !auto_accessor_instance_inits.is_empty())
-                || !auto_accessor_static_inits.is_empty()
-                || auto_accessor_class_alias.is_some())
-        {
-            self.write_line();
-            let mut wrote_alias_line = false;
-
-            if let Some(alias) = auto_accessor_class_alias.as_ref()
-                && !alias.is_empty()
-                && !class_name.is_empty()
-            {
-                self.write(alias);
-                self.write(" = ");
-                self.write(&class_name);
-                wrote_alias_line = true;
-            }
-
-            if emit_auto_accessor_instance_inits_after_class
-                && !auto_accessor_instance_inits.is_empty()
-            {
-                if wrote_alias_line {
-                    self.write(", ");
-                }
-                let mut wrote_instance_line = false;
-                for (i, (storage_name, _init_idx)) in
-                    auto_accessor_instance_inits.iter().enumerate()
-                {
-                    if i > 0 {
-                        self.write(", ");
-                    }
-                    self.write(storage_name);
-                    self.write(" = new WeakMap()");
-                    wrote_instance_line = true;
-                }
-                if wrote_alias_line || wrote_instance_line {
-                    self.write(";");
-                    self.write_line();
-                }
-            } else if wrote_alias_line {
-                self.write(";");
-                self.write_line();
-            }
-
-            for (storage_name, init_idx) in &auto_accessor_static_inits {
-                self.write(storage_name);
-                self.write(" = { value: ");
-                if let Some(init) = init_idx {
-                    self.emit_expression(*init);
-                } else {
-                    self.write("void 0");
-                }
-                self.write(" };");
-                self.write_line();
-            }
-        }
-
-        // For class expressions with private field lowering, emit the WeakMap/WeakSet/method
-        // initializations as comma-separated items inside the wrapping expression:
-        //   (_a = class C { ... },
-        //       _C_field = new WeakMap(),
-        //       _C_instances = new WeakSet(),
-        //       _C_method = function _C_method() { },
-        //       _a)
-        // For class declarations, emit as separate statements after the class body.
-        if needs_private_comma_expr
-            && has_post_class_inits
-            && !emitted_private_comma_items_before_static_items
-        {
-            self.emit_private_comma_items(PrivateCommaItems {
-                weakmap_inits: &weakmap_inits,
-                instances_ws: instances_ws.as_deref(),
-                private_auto_instance_storage_inits: &private_auto_instance_storage_inits,
-                method_defs: &method_defs,
-                accessor_defs: &accessor_defs,
-                private_member_def_needs_class_alias,
-                class_value_alias: class_value_alias.as_deref(),
-                class_name: &class_name,
-                emitted_private_auto_accessors_pre_static,
-                private_auto_accessors: &private_auto_accessors,
-                private_class_alias_pair: private_class_alias_pair.as_ref(),
-                set_function_name: if !needs_static_comma_expr
-                    || class_expr_static_comma_has_no_scheduled_elements
-                {
-                    class_expr_temp
-                        .as_deref()
-                        .zip(class_expr_set_function_name.as_deref())
-                } else {
-                    None
-                },
-                static_private_inits: &static_private_inits,
-            });
-
-            if !target_supports_native_private_names && has_legacy_private_name_member_decorators {
-                self.write(",");
-                self.write_line();
-                self.increase_indent();
-                self.write("(() => {");
-                self.write_line();
-                self.increase_indent();
-                self.emit_legacy_member_decorator_calls_requiring_private_name_scope(
-                    &class_name,
-                    &class.members.nodes,
-                );
-                self.decrease_indent();
-                self.write("})()");
-                self.decrease_indent();
-            }
-
-            // Close the comma expression with the temp var, unless the static field
-            // comma expr path will handle the closing.
-            if (!needs_static_comma_expr || class_expr_static_comma_has_no_scheduled_elements)
-                && let Some(ref temp) = class_expr_temp
-            {
-                self.write(",");
-                self.write_line();
-                self.increase_indent();
-                self.write(temp);
-                if class_expr_comma_needs_parens {
-                    self.write(")");
-                }
-                self.decrease_indent();
-                if assignment_prefix.is_some() {
-                    self.write(";");
-                }
-            }
-        } else if has_post_class_inits && !emitted_private_comma_items_before_static_items {
-            self.write_line();
-            let mut first = true;
-
-            // Class alias: _a = ClassName
-            if let Some((ref alias, ref cls_name)) = private_class_alias_pair {
-                self.write(alias);
-                self.write(" = ");
-                self.write(cls_name);
-                first = false;
-            }
-
-            // WeakMap inits first (tsc order): _X_field = new WeakMap()
-            for init in &weakmap_inits {
-                if !first {
-                    self.write(", ");
-                }
-                self.write(init);
-                first = false;
-            }
-
-            // WeakSet: _X_instances = new WeakSet()
-            if let Some(ref ws_name) = instances_ws {
-                if !first {
-                    self.write(", ");
-                }
-                self.write(ws_name);
-                self.write(" = new WeakSet()");
-                first = false;
-            }
-
-            for init in &private_auto_instance_storage_inits {
-                if !first {
-                    self.write(", ");
-                }
-                self.write(init);
-                first = false;
-            }
-
-            // Private method function definitions:
-            // _C_method = function _C_method(params) { ... }
-            for def in &method_defs {
-                if !first {
-                    self.write(", ");
-                }
-                self.emit_private_method_function_def(
-                    def,
-                    private_member_def_needs_class_alias,
-                    class_value_alias.as_deref(),
-                    &class_name,
-                );
-                first = false;
-            }
-
-            // Private accessor function definitions:
-            // _C_prop_get = function _C_prop_get() { ... }
-            // _C_prop_set = function _C_prop_set(param) { ... }
-            for def in &accessor_defs {
-                if !first {
-                    self.write(", ");
-                }
-                self.emit_private_accessor_function_def(
-                    def,
-                    private_member_def_needs_class_alias,
-                    class_value_alias.as_deref(),
-                    &class_name,
-                );
-                first = false;
-            }
-
-            if !emitted_private_auto_accessors_pre_static {
-                for accessor in &private_auto_accessors {
-                    if !first {
-                        self.write(", ");
-                    }
-                    self.emit_private_auto_accessor_function_def(
-                        &accessor.get_var_name,
-                        &accessor.storage_name,
-                        accessor.is_static,
-                        true,
-                        private_class_alias_pair
-                            .as_ref()
-                            .map(|(alias, _)| alias.as_str())
-                            .or(class_value_alias.as_deref()),
-                    );
-                    self.write(", ");
-                    self.emit_private_auto_accessor_function_def(
-                        &accessor.set_var_name,
-                        &accessor.storage_name,
-                        accessor.is_static,
-                        false,
-                        private_class_alias_pair
-                            .as_ref()
-                            .map(|(alias, _)| alias.as_str())
-                            .or(class_value_alias.as_deref()),
-                    );
-                    first = false;
-                }
-            }
-
-            self.write(";");
-        }
-
-        if !needs_private_comma_expr
-            && !target_supports_native_private_names
-            && has_legacy_private_name_member_decorators
-        {
-            if !self.writer.is_at_line_start() {
-                self.write_line();
-            }
-            self.write("(() => {");
-            self.write_line();
-            self.increase_indent();
-            self.emit_legacy_member_decorator_calls_requiring_private_name_scope(
-                &class_name,
-                &class.members.nodes,
-            );
-            self.decrease_indent();
-            self.write("})();");
-        }
-
-        // Emit static private field value initializations after class body:
-        // `_A_field = { value: 10 };`
-        // For class expressions with private lowering, these are already emitted
-        // as comma items above in the private comma expr block.
-        if needs_private_comma_expr {
-            // Already emitted above in the comma expression block.
-        } else {
-            for init in &static_private_inits {
-                self.write_line();
-                self.emit_static_private_init(init, &class_name, true);
-            }
-            for accessor in private_auto_accessors.iter().filter(|a| a.is_static) {
-                self.write_line();
-                self.write(&accessor.storage_name);
-                self.write(" = { value: ");
-                if let Some(init) = accessor.initializer {
-                    self.emit_expression(init);
-                } else {
-                    self.write("void 0");
-                }
-                self.write(" };");
-            }
-        }
-
-        // Emit deferred static blocks as IIFEs after the class body.
-        // Class expressions lowered to comma expressions must keep static block
-        // evaluation inside that expression before returning the temp.
-        if let Some(temp) = class_expr_static_temp.as_ref()
-            && static_field_inits.is_empty()
-            && !self.defer_class_static_blocks
-            && !deferred_static_blocks.is_empty()
-        {
-            if let Some(name) = class_expr_set_function_name.as_ref() {
-                self.emit_class_expr_set_function_name_comma_item(temp, name);
-            }
-            let prev_self_alias = self.scoped_class_expression_self_alias.clone();
-            if class_name_is_real && !class_name.is_empty() && class_name != *temp {
-                self.scoped_class_expression_self_alias = Some((
-                    Arc::<str>::from(class_name.as_str()),
-                    Arc::<str>::from(temp.as_str()),
-                ));
-            }
-            self.emit_static_block_iife_comma_items_with_context(
-                deferred_static_blocks,
-                static_initializer_this_binding,
-                static_initializer_super_base,
-            );
-            self.scoped_class_expression_self_alias = prev_self_alias;
-            self.write(",");
-            self.write_line();
-            self.increase_indent();
-            self.write(temp);
-            if class_expr_comma_needs_parens {
-                self.write(")");
-            }
-            self.decrease_indent();
-            if assignment_prefix.is_some() {
-                self.write(";");
-            }
-        } else if self.defer_class_static_blocks {
-            self.deferred_class_static_blocks
-                .extend(deferred_static_blocks);
-        } else {
-            if static_field_inits.is_empty()
-                && !deferred_static_blocks.is_empty()
-                && !emit_private_inits_before_static_elements
-                && !class_name.is_empty()
-                && let Some(alias) = static_initializer_class_alias.as_ref()
-            {
-                self.write_line();
-                self.write(alias);
-                self.write(" = ");
-                self.write(&class_name);
-                self.write(";");
-            }
-            let prev_self_alias = self.scoped_class_expression_self_alias.clone();
-            if let Some(alias) = static_initializer_class_alias.as_ref()
-                && !class_name.is_empty()
-            {
-                self.scoped_class_expression_self_alias = Some((
-                    Arc::<str>::from(class_name.as_str()),
-                    Arc::<str>::from(alias.as_str()),
-                ));
-            }
-            self.emit_static_block_iifes_with_context(
-                deferred_static_blocks,
-                static_initializer_this_binding,
-                static_initializer_super_base,
-            );
-            self.scoped_class_expression_self_alias = prev_self_alias;
-        }
+        self.emit_class_es6_after_body(ClassEs6AfterBody {
+            node,
+            class,
+            class_name: &class_name,
+            class_name_is_real,
+            assignment_prefix_is_some: assignment_prefix.is_some(),
+            static_initializer_self_alias,
+            static_initializer_this_binding,
+            static_initializer_super_base,
+            externalized_static_initializer_uses_undefined_receiver,
+            computed_side_effects_emitted_in_static_block,
+            class_expr_comma_needs_parens,
+            needs_computed_prop_comma_expr,
+            needs_static_comma_expr,
+            needs_private_comma_expr,
+            has_any_private_lowering,
+            private_member_def_needs_class_alias,
+            lower_auto_accessors_to_weakmap,
+            emit_auto_accessor_instance_inits_after_class,
+            target_supports_native_private_names,
+            has_legacy_private_name_member_decorators,
+            class_expr_temp,
+            class_expr_static_temp,
+            default_export_set_function_name_temp,
+            static_initializer_class_alias,
+            class_value_alias,
+            auto_accessor_class_alias,
+            class_expr_set_function_name,
+            computed_prop_entries,
+            computed_prop_entries_consumed_by_member_name,
+            computed_property_side_effects,
+            static_field_inits,
+            deferred_static_blocks,
+            private_auto_accessors: &private_auto_accessors,
+            auto_accessor_instance_inits: &auto_accessor_instance_inits,
+            auto_accessor_static_inits: &auto_accessor_static_inits,
+        });
 
         // Restore private field state (for nested classes)
         self.private_field_weakmaps = prev_private_field_weakmaps;
